switched UITableController to UIVewContr. + UITable
