+++ VK/VK/Control/AlphabetControl.swift
@@ -1,5 +1,5 @@
 //
-//  LikeControl.swift
+//  AlphabetControl.swift
 //  VK
 //
 //  Created by Lera on 26.06.22.
@@ -7,7 +7,8 @@
 
 import UIKit
 
-class LikeControl: UIControl {
-    
+class AlphabetControl: UIControl {
+
+
 
 }
